requirements.txt
2captcha_python==1.1.3
PyYAML==6.0
requests==2.28.1
twocaptcha==0.0.1
web3==5.31.3
websocket_client==1.4.2
